gpu: nvgpu: Fix kmem function definition
Use the correct type for the size argument in __nvgpu_kmalloc() and
make sure the size_t type is available for usage.

Bug 1799159
Bug 1823380

Reviewed-on: http://git-master/r/1318308
GVS: Gerrit_Virtual_Submit

+++ drivers/gpu/nvgpu/include/nvgpu/kmem_linux.h
@@ -52,7 +52,7 @@ static inline void nvgpu_kmem_debugfs_init(struct device *dev)
  * nvgpu. This should not be included directly - instead include <nvgpu/kmem.h>.
  */
 
-static inline void *__nvgpu_kmalloc(struct gk20a *g, unsigned long size,
+static inline void *__nvgpu_kmalloc(struct gk20a *g, size_t size,
 				    unsigned long ip)
 {
 #ifdef CONFIG_NVGPU_TRACK_MEM_USAGE
